refactor: 인증 설정을 lib/auth.ts로 분리 (관심사 분리)
